+++ changelog.md
@@ -1,3 +1,7 @@
+# 1.1.3.0
+
+* Fix malformed `signum` for `F2Poly`.
+
 # 1.1.2.0
 
 * Fix `setBit`, `clearBit`, `complementBit` to preserve vector's length.